Исправления (модальное окно удаления карточки)
