+++ README.md
@@ -7,5 +7,9 @@ grid-layout и др., верстка форм и полей, верстка мо
 ## Технологии
 HTML, CSS, ОфмфЫскшзеБЭМ-методология, flexbox, grid-layout
 
+## Разработчики
+Татьяна Ковпак | tanechka.kovpak@yandex.ru | https://github.com/TatianaKovpak
+Максим Акимов | instr-metodist@yandex.ru | https://github.com/maxim-akimov
+
 ## GitHub Pages
 https://maxim-akimov.github.io/mesto-project/
